Revert "Add tag in VCS support"
While useful in some cases, it isn't in the general case.

Plus, Bashi's tool fits together cleanly with versionah and histofile.
The single tools stacked together approach is far nicer in the instances
where versionah is currently used, so we'll go that way.

Support never hit a release, so there is no deprecation period.

This reverts commit 7209f0948dc405f445f5f0bccd967627c8057591.